+++ priv/repo/migrations/20221004064232_add_global_flag_to_user_search_filters.exs
@@ -1,0 +1,9 @@
+defmodule TdDd.Repo.Migrations.AddGlobalFlagToUserSearchFilters do
+  use Ecto.Migration
+
+  def change do
+    alter table("user_search_filters") do
+      add :is_global, :boolean, default: false
+    end
+  end
+end