+++ test/support/factory.ex
@@ -137,7 +137,7 @@ defmodule TdDd.Factory do
       domain_id: 2,
       dataset: build(:dataset),
       populations: build(:populations),
-      validations: build(:validations),
+      validation: build(:validation),
       segments: build(:segments),
       version: 1,
       status: :draft
@@ -168,7 +168,7 @@ defmodule TdDd.Factory do
       domain_id: 2,
       dataset: build(:dataset),
       populations: build(:populations),
-      validations: build(:validations),
+      validation: build(:validation),
       version: 1,
       status: :draft
     }
@@ -282,7 +282,8 @@ defmodule TdDd.Factory do
       name: sequence("filter_name"),
       filters: %{country: ["Sp"]},
       user_id: sequence(:user_id, & &1),
-      scope: :data_structure
+      scope: :data_structure,
+      is_global: false
     }
   end
 
@@ -393,14 +394,18 @@ defmodule TdDd.Factory do
 
   def populations_factory(_attrs) do
     [
-      %TdDq.Implementations.Populations{
-        population: [build(:condition_row)]
+      %TdDq.Implementations.Conditions{
+        conditions: [build(:condition_row)]
       }
     ]
   end
 
-  def validations_factory(_attrs) do
-    [build(:condition_row)]
+  def validation_factory(_attrs) do
+    [
+      %TdDq.Implementations.Conditions{
+        conditions: [build(:condition_row)]
+      }
+    ]
   end
 
   def segments_factory(_attrs) do